# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Service C - API Keanggotaan & Voucher",
        "description": "Service Smart Parking untuk mengelola data keanggotaan dan voucher parkir.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/api/v1/memberships": {
            "get": {
                "tags": [
                    "Keanggotaan"
                ],
                "summary": "Melihat daftar seluruh member",
                "operationId": "eaf4ebb907ef7726f6f19c8fdda52d5e",
                "responses": {
                    "200": {
                        "description": "Data berhasil diambil"
                    },
                    "401": {
                        "description": "Tidak terotorisasi"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Keanggotaan"
                ],
                "summary": "Mendaftarkan member baru",
                "operationId": "f3d5a078dd21f5146d6cddaf5838cc8e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "membership_type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Rani Putri"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "rani@mail.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "081234567899"
                                    },
                                    "membership_type": {
                                        "type": "string",
                                        "example": "perak",
                                        "enum": [
                                            "perunggu",
                                            "perak",
                                            "emas",
                                            "platina"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Anggota berhasil didaftarkan"
                    },
                    "400": {
                        "description": "Validasi gagal"
                    },
                    "401": {
                        "description": "Tidak terotorisasi"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/memberships/{id}": {
            "get": {
                "tags": [
                    "Keanggotaan"
                ],
                "summary": "Mengecek detail dan status aktif seorang member",
                "operationId": "cb9af727af4ffaafb8d1edd379452f00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Member code (contoh: MEM001)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "MEM001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil diambil"
                    },
                    "404": {
                        "description": "Anggota tidak ditemukan"
                    },
                    "401": {
                        "description": "Tidak terotorisasi"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/vouchers/{code}": {
            "get": {
                "tags": [
                    "Voucher"
                ],
                "summary": "Melihat detail voucher",
                "operationId": "e35ffc2d970416bc1abd54e7ffc08451",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Kode voucher",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "PARKIRHEMAT"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil diambil"
                    },
                    "404": {
                        "description": "Voucher tidak ditemukan"
                    },
                    "401": {
                        "description": "Tidak terotorisasi"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/vouchers/{code}/use": {
            "patch": {
                "tags": [
                    "Voucher"
                ],
                "summary": "Menandai voucher sebagai sudah terpakai",
                "operationId": "72a7cf4ecdeddf4dcf44a17693b07c62",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Kode voucher",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "PARKIRHEMAT"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Voucher berhasil ditandai sebagai terpakai"
                    },
                    "404": {
                        "description": "Voucher tidak ditemukan"
                    },
                    "400": {
                        "description": "Voucher sudah terpakai"
                    },
                    "401": {
                        "description": "Tidak terotorisasi"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Membership": {
                "properties": {
                    "id": {
                        "description": "Base controller.",
                        "type": "integer",
                        "example": 1
                    },
                    "member_code": {
                        "type": "string",
                        "example": "MEM001"
                    },
                    "name": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "email": {
                        "type": "string",
                        "example": "budi@mail.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "081234567890"
                    },
                    "membership_type": {
                        "type": "string",
                        "example": "emas",
                        "enum": [
                            "perunggu",
                            "perak",
                            "emas",
                            "platina"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "example": "aktif",
                        "enum": [
                            "aktif",
                            "kedaluwarsa"
                        ]
                    },
                    "discount_percent": {
                        "type": "integer",
                        "example": 20
                    },
                    "registered_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "expired_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "API Key (masukkan NIM Anda) untuk otorisasi",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Keanggotaan",
            "description": "Keanggotaan"
        },
        {
            "name": "Voucher",
            "description": "Voucher"
        }
    ]
}